Multi POD & Lights
Program now allows for multiple StageKits to be connected to a single RPi.  Also data pass-through to StageKits has been enabled.
WARNING: Old INI files will not work, please use included INI file & change as you require.

+++ StageKitPied/stagekitpied.cpp
@@ -48,14 +48,14 @@ int main(int arc, char *argv[]) {
   MSG_SKP_INFO( "Started" );
 
   // Sleepy timer
-  uint16_t sleepyTime = 0;
+  long sleepyTime = 0;
 
   SleepTimer sleeper;
   sleeper.SetSleepTimeMax( sleepyTime );
   sleeper.Start();
 
-  uint16_t slept_ms = 10;
-  uint16_t sleep_time_new = 10;
+  long slept_ms = 10;
+  long sleep_time_new = 10;
 
   ConsoleInput console;
